Releasing code into production

+++ MoneyFi-backend/MoneyFi-budget/src/main/java/com/moneyfi/budget/service/impl/BudgetServiceImpl.java
@@ -1,6 +1,7 @@
 package com.moneyfi.budget.service.impl;
 
 import com.moneyfi.budget.exceptions.ResourceNotFoundException;
+import com.moneyfi.budget.exceptions.ScenarioNotPossibleException;
 import com.moneyfi.budget.model.BudgetModel;
 import com.moneyfi.budget.repository.BudgetRepository;
 import com.moneyfi.budget.repository.common.BudgetCommonRepository;
@@ -9,6 +10,7 @@ import com.moneyfi.budget.service.dto.request.AddBudgetDto;
 import com.moneyfi.budget.service.dto.response.BudgetDetailsDto;
 import com.moneyfi.budget.service.dto.response.SpendingAnalysisResponseDto;
 import com.moneyfi.budget.service.dto.response.UserDetailsForSpendingAnalysisDto;
+import com.moneyfi.budget.utils.BudgetValidator;
 import com.moneyfi.budget.utils.GeneratePdfTemplate;
 import com.moneyfi.budget.utils.StringConstants;
 import jakarta.transaction.Transactional;
@@ -21,10 +23,10 @@ import org.springframework.web.client.RestTemplate;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.*;
 
-import static com.moneyfi.budget.utils.StringConstants.BUDGET_NOT_FOUND;
-
+import static com.moneyfi.budget.utils.StringConstants.*;
 
 @Service
 public class BudgetServiceImpl implements BudgetService {
@@ -45,8 +47,13 @@ public class BudgetServiceImpl implements BudgetService {
     }
 
     @Override
-    @Transactional
+    @Transactional(rollbackOn = Exception.class)
     public void saveBudget(List<AddBudgetDto> budgetList, Long userId) {
+        Optional<List<BudgetModel>> existingBudget = budgetRepository.findByUserId(userId);
+        if(existingBudget.isPresent() && !existingBudget.get().isEmpty()) {
+            throw new ScenarioNotPossibleException("Budget already exists! Please update if required");
+        }
+        BudgetValidator.validateBudgetSaveRequestDto(budgetList, getTotalIncomeInMonthAndYear(userId, LocalDateTime.now().getMonthValue(), LocalDateTime.now().getYear()));
         List<BudgetModel> newBudget = new ArrayList<>();
         for (AddBudgetDto budget : budgetList) {
             BudgetModel budgetModel = new BudgetModel();
@@ -64,32 +71,24 @@ public class BudgetServiceImpl implements BudgetService {
 
     @Override
     public BigDecimal budgetProgress(Long userId, int month, int year) {
-        List<BudgetDetailsDto> budgetsList = getAllBudgetsByUserIdAndCategory(userId, month, year, "all");
-        BigDecimal moneyLimit = budgetsList
-                            .stream()
-                            .map(BudgetDetailsDto::getMoneyLimit)
-                            .reduce(BigDecimal.ZERO, BigDecimal::add);
-        BigDecimal currentSpending = getTotalExpenseInMonthAndYear(userId, month, year);
-        return currentSpending.divide(moneyLimit, 5, RoundingMode.HALF_UP);
-    }
-
-    private BigDecimal getTotalExpenseInMonthAndYear(Long userId, int month, int year) {
-        BigDecimal totalExpense = budgetRepository.getTotalExpenseInMonthAndYear(userId, month, year);
-        if(totalExpense == null){
-            return BigDecimal.ZERO;
-        }
-        return totalExpense;
+        return getTotalExpenseInMonthAndYear(userId, month, year).divide(getAllBudgetsByUserIdAndCategory(userId, month, year, "all")
+                .stream()
+                .map(BudgetDetailsDto::getMoneyLimit)
+                .reduce(BigDecimal.ZERO, BigDecimal::add), 5, RoundingMode.HALF_UP);
     }
 
     @Override
-    @Transactional
+    @Transactional(rollbackOn = Exception.class)
     public void updateBudget(Long userId, List<BudgetModel> budgetList) {
+        BudgetValidator.validateBudgetUpdateRequestDto(budgetList);
         List<BudgetModel> budgetListToUpdate = new ArrayList<>();
+        LocalDateTime currentTime = LocalDateTime.now();
         for (BudgetModel budget : budgetList) {
             BudgetModel budgetModel = budgetRepository.findById(budget.getId()).orElseThrow(() -> new ResourceNotFoundException(BUDGET_NOT_FOUND));
             if (budget.getMoneyLimit().compareTo(BigDecimal.ZERO) >= 0) {
                 budgetModel.setMoneyLimit(budget.getMoneyLimit());
             }
+            budgetModel.setUpdatedAt(currentTime);
             budgetListToUpdate.add(budgetModel);
         }
         budgetRepository.saveAll(budgetListToUpdate);
@@ -172,23 +171,38 @@ public class BudgetServiceImpl implements BudgetService {
         try {
             byte[] pdfBytes = getUserSpendingAnalysisByBudgetCategoriesPdf(userId, fromDate, toDate, authHeader);
             apiCallToGatewayServiceToSendEmail(pdfBytes, authHeader);
-            return ResponseEntity.ok("Email sent successfully");
+            return ResponseEntity.ok(EMAIL_SENT_SUCCESS_MESSAGE);
         } catch (Exception e) {
             e.printStackTrace();
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body("Failed to send email: " + e.getMessage());
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(EMAIL_SENT_FAILURE_MESSAGE + ": " + e.getMessage());
         }
     }
 
     @Override
-    @Transactional
+    @Transactional(rollbackOn = Exception.class)
     public void deleteBudget(Long userId) {
         Optional<List<BudgetModel>> budgetList = budgetRepository.findByUserId(userId);
-        if (budgetList.isPresent()) {
+        if (budgetList.isPresent() && !budgetList.get().isEmpty()) {
             budgetRepository.deleteAll(budgetList.get());
         } else {
             throw new ResourceNotFoundException(BUDGET_NOT_FOUND);
         }
+    }
+
+    private BigDecimal getTotalIncomeInMonthAndYear(Long userId, int month, int year) {
+        BigDecimal totalIncome = budgetRepository.getTotalIncomeInMonthAndYear(userId, month, year);
+        if(totalIncome == null){
+            return BigDecimal.ZERO;
+        }
+        return totalIncome;
+    }
+
+    private BigDecimal getTotalExpenseInMonthAndYear(Long userId, int month, int year) {
+        BigDecimal totalExpense = budgetRepository.getTotalExpenseInMonthAndYear(userId, month, year);
+        if(totalExpense == null){
+            return BigDecimal.ZERO;
+        }
+        return totalExpense;
     }
 
     private void apiCallToGatewayServiceToSendEmail(byte[] pdfBytes, String authHeader){
@@ -199,24 +213,13 @@ public class BudgetServiceImpl implements BudgetService {
 
         HttpEntity<byte[]> requestEntity = new HttpEntity<>(pdfBytes, headers);
         ResponseEntity<Void> response = restTemplate.exchange(
-                StringConstants.API_GATEWAY_URL_PROFILE_CONTROLLER + "spending-analysis/email",
+                StringConstants.USER_SERVICE_URL_CONTROLLER + "/spending-analysis/email",
                 HttpMethod.POST,
                 requestEntity,
                 Void.class
         );
         if (!response.getStatusCode().is2xxSuccessful()) {
-            throw new ResourceNotFoundException("Failed to send email: " + response.getStatusCode());
+            throw new ResourceNotFoundException(EMAIL_SENT_FAILURE_MESSAGE + ": " + response.getStatusCode());
         }
-    }
-
-    private String makeUsernamePrivate(String username){
-        int index = username.indexOf('@');
-        return username.substring(0, index/3) +
-                "x".repeat(index - index/3) + username.substring(index);
-    }
-
-    private String generateDocumentPasswordForUser(UserDetailsForSpendingAnalysisDto userDetails){
-        return userDetails.getName().substring(0,4).toUpperCase() +
-                userDetails.getUsername().substring(0,4).toLowerCase();
     }
 }
